+++ src/api/objects/toot.ts
@@ -565,7 +565,7 @@ export default class Toot implements TootObj {
         const tootsToComplete = toots.filter(toot => toot instanceof Toot ? toot.shouldComplete() : true);
 
         const batchSleepMS = isDeepInspect ? Config.sleepBetweenCompletionMS : 0;
-        const newCompleteToots = await batchMap(tootsToComplete, (t) => complete(t), "completeToots", null, batchSleepMS);
+        const newCompleteToots = await batchMap(tootsToComplete, (t) => complete(t), "completeToots", 50, batchSleepMS);
         toots = completeToots.concat(newCompleteToots);
 
         let msg = `${logPrefix} completeToots(isDeepInspect=${isDeepInspect}) on ${toots.length} toots`;