+++ src/config.ts
@@ -128,7 +128,7 @@ export const Config: ConfigType = {
     mutexWarnSeconds: 5,                    // How long to wait before warning about a mutex lock
     numServersToCheck: 30,                  // NUM_SERVERS_TO_CHECK
     reloadFeaturesEveryNthOpen: 9,          // RELOAD_FEATURES_EVERY_NTH_OPEN
-    sleepBetweenCompletionMS: 200,          // How long to wait between batches of Toot.completeToots() calls
+    sleepBetweenCompletionMS: 250,          // How long to wait between batches of Toot.completeToots() calls
     timeoutMS: 5_000,                       // Timeout for API calls
 
     // Trending tags and links