+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* call `Toot.complete()` in batches of 50 instead of 100, increase sleep time.
 
 # v0.30.0
 * (Demo App) Implement viewing and voting in Mastodon polls