+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+
+### v0.30.1
 * Call `Toot.complete()` in batches of 50 instead of 100, increase sleep time to 250ms from 200
 * Don't penalize toots from followed accounts and tags with `TRENDING` modifier if it's less than 1 (and thus a penalty)
 * Apply the `OUTLIER_DAMPENER` to the `DiversityFeedScorer` penalty